refactor: port extension pull and rm to libswamp generator + renderer pattern (#739)
Move domain logic for `extension pull` and `extension rm` out of CLI
commands into libswamp event-yielding generators with testable deps
injection, following the established pattern from model delete, data gc,
and extension push.

**extension rm** (new: src/libswamp/extensions/rm.ts)
- `extensionRmPreview()` — async preview with validation
- `extensionRm()` — async generator yielding deleting → completed
- `ExtensionRmDeps` with plain function signatures (no infrastructure leaks)
- `createExtensionRmDeps()` factory wiring real FS operations
- Local `UpstreamMap`/`UpstreamEntry` types replacing infrastructure type

**extension pull** (new: src/libswamp/extensions/pull.ts)
- `extensionPull()` — async generator wrapping `installExtension()`
- `ExtensionPullDeps` with function signatures for registry API calls
  (`getExtension`, `downloadArchive`, `getChecksum`) instead of
  exposing `ExtensionApiClient`
- `createExtensionPullDeps()` / `createInstallContext()` factories
  instantiate infrastructure internally
- Local `ExtensionSafetyWarning` and `ExtensionRegistryInfo` types
- `ConflictError` propagates to CLI for two-phase prompt flow
- All moved domain helpers: `parseExtensionRef`, `validateExtensionName`,
  `installExtension`, `detectConflicts`, conflict detection, symlink
  validation, source completeness checks

**Renderers** (new: src/presentation/renderers/)
- `ExtensionPullRenderer` with `renderConflicts()` extension method
- `ExtensionRmRenderer` with `renderDependencyWarning()` extension method
- Log and JSON variants matching original output exactly

**CLI commands** slimmed to orchestration:
- Create LibSwampContext + deps + renderer
- preview → prompt → consumeStream(generator, renderer.handlers())
- Conflict-retry flow preserved for pull

**Standards compliance:**
- CLI commands and renderers import only from `libswamp/mod.ts`
- No infrastructure types in deps interfaces
- Factory functions encapsulate infrastructure instantiation
- All event types have `completed` + `error` terminals

Closes #739 (Batch 5)
