+++ src/libswamp/extensions/pull_test.ts
@@ -1,0 +1,99 @@
+// Swamp, an Automation Framework
+// Copyright (C) 2026 System Initiative, Inc.
+//
+// This file is part of Swamp.
+//
+// Swamp is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License version 3
+// as published by the Free Software Foundation, with the Swamp
+// Extension and Definition Exception (found in the "COPYING-EXCEPTION"
+// file).
+//
+// Swamp is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Swamp.  If not, see <https://www.gnu.org/licenses/>.
+
+import { assertEquals, assertThrows } from "@std/assert";
+import { assertStringIncludes } from "@std/assert/string-includes";
+import {
+  parseExtensionRef,
+  updateUpstreamExtensions,
+  validateExtensionName,
+} from "./pull.ts";
+import { UserError } from "../../domain/errors.ts";
+
+Deno.test("parseExtensionRef: parses name without version", () => {
+  const ref = parseExtensionRef("@myorg/my-ext");
+  assertEquals(ref.name, "@myorg/my-ext");
+  assertEquals(ref.version, null);
+});
+
+Deno.test("parseExtensionRef: parses name with version", () => {
+  const ref = parseExtensionRef("@myorg/my-ext@2026.02.26.1");
+  assertEquals(ref.name, "@myorg/my-ext");
+  assertEquals(ref.version, "2026.02.26.1");
+});
+
+Deno.test("parseExtensionRef: throws on missing @ prefix", () => {
+  assertThrows(
+    () => parseExtensionRef("myorg/my-ext"),
+    UserError,
+    'must start with "@"',
+  );
+});
+
+Deno.test("parseExtensionRef: throws on empty version", () => {
+  assertThrows(
+    () => parseExtensionRef("@myorg/my-ext@"),
+    UserError,
+    "Version cannot be empty",
+  );
+});
+
+Deno.test("parseExtensionRef: parses nested segments", () => {
+  const ref = parseExtensionRef("@myorg/my-ext/sub");
+  assertEquals(ref.name, "@myorg/my-ext/sub");
+  assertEquals(ref.version, null);
+});
+
+Deno.test("validateExtensionName: accepts valid names", () => {
+  validateExtensionName("@myorg/my-ext");
+  validateExtensionName("@my_org/my_ext");
+  validateExtensionName("@myorg/my-ext/sub");
+});
+
+Deno.test("validateExtensionName: rejects invalid names", () => {
+  assertThrows(
+    () => validateExtensionName("myorg/my-ext"),
+    UserError,
+    "Must match",
+  );
+  assertThrows(
+    () => validateExtensionName("@MyOrg/My-Ext"),
+    UserError,
+    "Must match",
+  );
+});
+
+Deno.test("updateUpstreamExtensions: writes and updates entries", async () => {
+  const tmpDir = await Deno.makeTempDir({ prefix: "swamp_test_" });
+  try {
+    await updateUpstreamExtensions(tmpDir, "@test/first", "1.0.0", [
+      "a.yaml",
+    ]);
+
+    const content = await Deno.readTextFile(
+      `${tmpDir}/upstream_extensions.json`,
+    );
+    const data = JSON.parse(content);
+    assertEquals(data["@test/first"].version, "1.0.0");
+    assertEquals(data["@test/first"].files, ["a.yaml"]);
+    assertStringIncludes(data["@test/first"].pulledAt, "20");
+  } finally {
+    await Deno.remove(tmpDir, { recursive: true });
+  }
+});
